runtime/src/main/java/org/corfudb/runtime/clients/NettyClientRouter.java
package org.corfudb.runtime.clients;

import com.codahale.metrics.Counter;
import com.codahale.metrics.MetricRegistry;
import com.codahale.metrics.Timer;
import com.google.common.util.concurrent.ThreadFactoryBuilder;
import io.netty.bootstrap.Bootstrap;
import io.netty.channel.Channel;
import io.netty.channel.ChannelFuture;
import io.netty.channel.ChannelHandler;
import io.netty.channel.ChannelHandlerContext;
import io.netty.channel.ChannelInitializer;
import io.netty.channel.ChannelOption;
import io.netty.channel.DefaultChannelPromise;
import io.netty.channel.EventLoopGroup;
import io.netty.channel.SimpleChannelInboundHandler;
import io.netty.handler.codec.LengthFieldBasedFrameDecoder;
import io.netty.handler.codec.LengthFieldPrepender;
import io.netty.handler.ssl.SslContext;
import io.netty.util.concurrent.GlobalEventExecutor;
import java.time.Duration;
import java.util.ArrayList;
import java.util.List;
import java.util.Map;
import java.util.NoSuchElementException;
import java.util.concurrent.CompletableFuture;
import java.util.concurrent.ConcurrentHashMap;
import java.util.concurrent.ExecutionException;
import java.util.concurrent.TimeUnit;
import java.util.concurrent.TimeoutException;
import java.util.concurrent.atomic.AtomicLong;
import javax.annotation.Nonnull;
import javax.net.ssl.SSLException;
import lombok.Getter;
import lombok.Setter;
import lombok.extern.slf4j.Slf4j;
import org.corfudb.protocols.wireprotocol.ClientHandshakeHandler;
import org.corfudb.protocols.wireprotocol.ClientHandshakeHandler.ClientHandshakeEvent;
import org.corfudb.protocols.wireprotocol.CorfuMsg;
import org.corfudb.protocols.wireprotocol.CorfuMsgType;
import org.corfudb.protocols.wireprotocol.NettyCorfuMessageDecoder;
import org.corfudb.protocols.wireprotocol.NettyCorfuMessageEncoder;
import org.corfudb.runtime.CorfuRuntime;
import org.corfudb.runtime.CorfuRuntime.CorfuRuntimeParameters;
import org.corfudb.runtime.exceptions.NetworkException;
import org.corfudb.runtime.exceptions.ShutdownException;
import org.corfudb.runtime.exceptions.WrongEpochException;
import org.corfudb.runtime.exceptions.unrecoverable.UnrecoverableCorfuError;
import org.corfudb.runtime.exceptions.unrecoverable.UnrecoverableCorfuInterruptedError;
import org.corfudb.security.sasl.SaslUtils;
import org.corfudb.security.sasl.plaintext.PlainTextSaslNettyClient;
import org.corfudb.security.tls.SslContextConstructor;
import org.corfudb.util.CFUtils;
import org.corfudb.util.MetricsUtils;
import org.corfudb.util.NodeLocator;
import org.corfudb.util.Sleep;


/**
 * A client router which multiplexes operations over the Netty transport.
 *
 * <p>Created by mwei on 12/8/15.
 */
@Slf4j
@ChannelHandler.Sharable
public class NettyClientRouter extends SimpleChannelInboundHandler<CorfuMsg>
        implements IClientRouter {

    /**
     * Metrics: meter (counter), histogram.
     */
    private Timer timerConnect;
    private Timer timerSyncOp;
    private Counter counterSendDisconnected;

    /**
     * The epoch this router is in.
     */
    @Getter
    public long epoch;

    /**
     * We should never set epoch backwards.
     *
     * @param epoch
     */
    public void setEpoch(long epoch){
        if (epoch < this.epoch) {
            log.warn("setEpoch: Rejected attempt to set the "
                    + "router {} to epoch {} smaller than current epoch {}",
                node, epoch, this.epoch);
            return;
        }
        this.epoch = epoch;
    }

    /**
     * New connection timeout (milliseconds).
     */
    @Getter
    @Setter
    public long timeoutConnect;

    /**
     * Sync call response timeout (milliseconds).
     */
    @Getter
    @Setter
    public long timeoutResponse;

    /**
     * Retry interval after timeout (milliseconds).
     */
    @Getter
    @Setter
    public long timeoutRetry;

    /**
     * The current request ID.
     */
    @Getter
    @Deprecated // TODO: Add replacement method that conforms to style
    @SuppressWarnings("checkstyle:abbreviation") // Due to deprecation
    public AtomicLong requestID;
    /**
     * The handlers registered to this router.
     */
    public Map<CorfuMsgType, IClient> handlerMap;
    /**
     * The clients registered to this router.
     */
    public List<IClient> clientList;
    /**
     * The outstanding requests on this router.
     */
    public Map<Long, CompletableFuture> outstandingRequests;
    /**
     * The currently registered channel context.
     */
    public ChannelHandlerContext context;

    /**
     * The currently registered channel.
     */
    public volatile Channel channel = null;

    /**
     * The {@link EventLoopGroup} for this router which services requests.
     */
    public final EventLoopGroup eventLoopGroup;

    /** Whether to shutdown the {@code eventLoopGroup} or not. Only applies when
     *  a deprecated constructor (which generates its own {@link EventLoopGroup} is used.
     */
    private boolean shutdownEventLoop = false;

    /**
     * Whether or not this router is shutdown.
     */
    public volatile boolean shutdown;

    /** The {@link NodeLocator} which represents the remote node this
     *  {@link NettyClientRouter} connects to.
     */
    @Getter
    private final NodeLocator node;

    /** The {@link CorfuRuntimeParameters} used to configure the
     *  router.
     */
    private final CorfuRuntimeParameters parameters;

    /** A {@link CompletableFuture} which is completed when a connection,
     *  including a successful handshake completes and messages can be sent
     *  to the remote node.
     */
    @Getter
    volatile CompletableFuture<Void> connectionFuture;


    private SslContext sslContext;

    /**
     * Creates a new NettyClientRouter connected to the specified host and port with the
     * specified tls and sasl options. The new {@link this} will attempt connection to
     * the node until {@link this#stop()} is called.
     *
     * @param node           The node to connect to.
     * @param eventLoopGroup The {@link EventLoopGroup} for servicing I/O.
     * @param parameters     A {@link CorfuRuntimeParameters} with the desired configuration.
     */
    public NettyClientRouter(@Nonnull NodeLocator node,
        @Nonnull EventLoopGroup eventLoopGroup,
        @Nonnull CorfuRuntimeParameters parameters) {
        this.node = node;
        this.parameters = parameters;
        this.eventLoopGroup = eventLoopGroup;

        timeoutConnect = parameters.getConnectionTimeout().toMillis();
        timeoutResponse = parameters.getRequestTimeout().toMillis();
        timeoutRetry = parameters.getConnectionRetryRate().toMillis();

        connectionFuture = new CompletableFuture<>();

        handlerMap = new ConcurrentHashMap<>();
        clientList = new ArrayList<>();
        requestID = new AtomicLong();
        outstandingRequests = new ConcurrentHashMap<>();
        shutdown = true;

        MetricRegistry metrics = CorfuRuntime.getDefaultMetrics();

        String pfx = CorfuRuntime.getMpCR() + node + ".";
        timerConnect = metrics.timer(pfx + "connectAsync");
        timerSyncOp = metrics.timer(pfx + "sync-op");
        counterSendDisconnected = metrics.counter(pfx + "send-disconnected");

        if (parameters.isTlsEnabled()) {
            try {
                sslContext = SslContextConstructor.constructSslContext(false,
                    parameters.getKeyStore(),
                    parameters.getKsPasswordFile(),
                    parameters.getTrustStore(),
                    parameters.getTsPasswordFile());
            } catch (SSLException e) {
                throw new UnrecoverableCorfuError(e);
            }
        }

        addClient(new BaseClient());


        // Initialize the channel
        shutdown = false;
        Bootstrap b = new Bootstrap();
        b.group(eventLoopGroup);
        b.channel(parameters.getSocketType().getChannelClass());
        parameters.getNettyChannelOptions().forEach(b::option);
        b.handler(getChannelInitializer());
        b.option(ChannelOption.CONNECT_TIMEOUT_MILLIS,
                (int) parameters.getConnectionTimeout().toMillis());

        // Asynchronously connect, retrying until shut down.
        // Once connected, connectionFuture will be completed.
        connectAsync(b);
    }

    /**
     * Add a new client to the router.
     *
     * @param client The client to add to the router.
     * @return This NettyClientRouter, to support chaining and the builder pattern.
     */
    public IClientRouter addClient(IClient client) {
        // Set the client's router to this instance.
        client.setRouter(this);

        // Iterate through all types of CorfuMsgType, registering the handler
        client.getHandledTypes().stream()
            .forEach(x -> {
                handlerMap.put(x, client);
                log.trace("Registered {} to handle messages of type {}", client, x);
            });

        // Register this type
        clientList.add(client);
        return this;
    }

    /**
     * Gets a client that matches a particular type.
     *
     * @param clientType The class of the client to match.
     * @param <T>        The type of the client to match.
     * @return The first client that matches that type.
     * @throws NoSuchElementException If there are no clients matching that type.
     */
    @SuppressWarnings("unchecked")
    public <T extends IClient> T getClient(Class<T> clientType) {
        return (T) clientList.stream()
            .filter(clientType::isInstance)
            .findFirst().get();
    }

    /**
     * {@inheritDoc}.
     *
     * @deprecated The router automatically starts now, so this function call is no
     *             longer necessary
     */
    @Override
    @Deprecated
    public synchronized void start() {
        // Do nothing, legacy call
    }

    /** Get the {@link ChannelInitializer} used for initializing the Netty channel pipeline.
     *
     * @return A {@link ChannelInitializer} which initializes the pipeline.
     */
    private ChannelInitializer getChannelInitializer() {
        return new ChannelInitializer() {
            @Override
            protected void initChannel(@Nonnull Channel ch) throws Exception {
                if (parameters.isTlsEnabled()) {
                    ch.pipeline().addLast("ssl", sslContext.newHandler(ch.alloc()));
                }
                ch.pipeline().addLast(new LengthFieldPrepender(4));
                ch.pipeline().addLast(new LengthFieldBasedFrameDecoder(Integer.MAX_VALUE,
                    0, 4, 0,
                    4));
                if (parameters.isSaslPlainTextEnabled()) {
                    PlainTextSaslNettyClient saslNettyClient =
                        SaslUtils.enableSaslPlainText(parameters.getUsernameFile(),
                            parameters.getPasswordFile());
                    ch.pipeline().addLast("sasl/plain-text", saslNettyClient);
                }
                ch.pipeline().addLast(new NettyCorfuMessageDecoder());
                ch.pipeline().addLast(new NettyCorfuMessageEncoder());
                ch.pipeline().addLast(new ClientHandshakeHandler(parameters.getClientId(),
                    node.getNodeId(), parameters.getHandshakeTimeout()));
                ch.pipeline().addLast(NettyClientRouter.this);
            }
        };
    }

    /** Add a future which reconnects the server.
     *
     * @param channel       The channel to use
     * @param bootstrap     The channel bootstrap to use
     */
    private void addReconnectionOnCloseFuture(@Nonnull Channel channel,
            @Nonnull Bootstrap bootstrap) {
        channel.closeFuture().addListener((r) -> {
            log.info("addReconnectionOnCloseFuture[{}]: disconnected", node);
            // Remove the current completion future, forcing clients to wait for reconnection.
            connectionFuture = new CompletableFuture<>();
            // Exceptionally complete all requests that were waiting for a completion.
            outstandingRequests.forEach((reqId, reqCompletableFuture) -> {
                reqCompletableFuture.completeExceptionally(
                        new NetworkException("Disconnected", node));
            // And also remove them.
            outstandingRequests.remove(reqId);
            });
            // If we aren't shutdown, reconnect.
            if (!shutdown) {
                log.info("addReconnectionOnCloseFuture[{}]: reconnecting", node);
                // Asynchronously connect again.
                connectAsync(bootstrap);
            }
        });
    }

    /** Connect to a remote server asynchronously.
     *
     * @param bootstrap         The channel boostrap to use
     * @return                  A {@link ChannelFuture} which is c
     */
    private ChannelFuture connectAsync(@Nonnull Bootstrap bootstrap) {
        // If shutdown, return a ChannelFuture that is exceptionally completed.
        if (shutdown) {
            return new DefaultChannelPromise(channel, GlobalEventExecutor.INSTANCE)
                .setFailure(new ShutdownException("Runtime already shutdown!"));
        }
        // Use the bootstrap to create a new channel.
        ChannelFuture f = bootstrap.connect(node.getHost(), node.getPort());
        f.addListener((ChannelFuture cf) -> channelConnectionFutureHandler(cf, bootstrap));
        return f;
    }

    /** Handle when a channel is connected.
     *
     * @param future        The future that is completed when the channel is connected/
     * @param bootstrap     The bootstrap to connect a new channel (used on reconnect).
     */
    private void channelConnectionFutureHandler(@Nonnull ChannelFuture future,
                                                @Nonnull Bootstrap bootstrap) {
        if (future.isSuccess()) {
            // When the channel connection succeeds, set this channel as active
            channel = future.channel();
            // And register a future to reconnect in case we get disconnected
            addReconnectionOnCloseFuture(channel, bootstrap);
            log.info("connectAsync[{}]: Channel connected.", node);
        } else {
            // Otherwise, the connection failed. If we're not shutdown, try reconnecting after
            // a sleep period.
            if (!shutdown) {
                log.info("connectAsync[{}]: Channel connection failed, reconnecting...", node);
                Sleep.sleepUninterruptibly(parameters.getConnectionRetryRate());
                // Call connect, which will retry the call again.
                // Note that this is not recursive, because it is called in the
                // context of the handler future.
                connectAsync(bootstrap);
            }
        }
    }

    /**
     * Stops routing requests.
     */
    @Override
    public void stop() {
        log.debug("stop: Shutting down router for {}", node);
        shutdown = true;
        connectionFuture.completeExceptionally(new ShutdownException());
        try {
            channel.disconnect();
            channel.close().syncUninterruptibly();
        } catch (Exception e) {
            log.error("Error in closing channel");
        }
        try {
            if (shutdownEventLoop) {
                eventLoopGroup.shutdownGracefully().sync();
            }
        } catch (InterruptedException e) {
            throw new UnrecoverableCorfuInterruptedError("Interrupted while stopping", e);
        }
    }

    /** {@inheritDoc}
     *  @deprecated  Deprecated, stopping a router without shutting it down is no longer supported.
     *               Please use {@link this#stop()}.
     */
    @Override
    @Deprecated
    public void stop(boolean shutdown) {
        stop();
    }

    /**
     * Send a message and get a completable future to be fulfilled by the reply.
     *
     * @param ctx     The channel handler context to send the message under.
     * @param message The message to send.
     * @param <T>     The type of completable to return.
     * @return A completable future which will be fulfilled by the reply,
     *     or a timeout in the case there is no response.
     */
    public <T> CompletableFuture<T> sendMessageAndGetCompletable(ChannelHandlerContext ctx,
        CorfuMsg message) {
        boolean isEnabled = MetricsUtils.isMetricsCollectionEnabled();
        // Check the connection future, and wait for it to be successful before
        // sending the message.
        try {
            connectionFuture
                .get(parameters.getConnectionTimeout().toMillis(), TimeUnit.MILLISECONDS);
        } catch (InterruptedException e) {
            Thread.currentThread().interrupt();
            throw new UnrecoverableCorfuInterruptedError(e);
        } catch (TimeoutException | ExecutionException e) {
            // If we timed out, return a CompletableFuture exceptionally completed
            // with the timeout.
            CompletableFuture<T> f = new CompletableFuture<>();
            f.completeExceptionally(e);
            return f;
        }
        final Timer.Context context = MetricsUtils
            .getConditionalContext(isEnabled, timerSyncOp);
        // Get the next request ID.
        final long thisRequest = requestID.getAndIncrement();
        // Set the message fields.
        message.setClientID(parameters.getClientId());
        message.setRequestID(thisRequest);
        message.setEpoch(epoch);

        // Generate a future and put it in the completion table.
        final CompletableFuture<T> cf = new CompletableFuture<>();
        outstandingRequests.put(thisRequest, cf);
        // Write the message out to the channel.
        if (ctx == null) {
            channel.writeAndFlush(message, channel.voidPromise());
        } else {
            ctx.writeAndFlush(message, ctx.voidPromise());
        }
        log.trace("Sent message: {}", message);
        final CompletableFuture<T> cfElapsed = cf.thenApply(x -> {
            MetricsUtils.stopConditionalContext(context);
            return x;
        });
        // Generate a timeout future, which will complete exceptionally
        // if the main future is not completed.
        final CompletableFuture<T> cfTimeout =
            CFUtils.within(cfElapsed, Duration.ofMillis(timeoutResponse));
        cfTimeout.exceptionally(e -> {
            outstandingRequests.remove(thisRequest);
            log.debug("Remove request {} due to timeout!", thisRequest);
            return null;
        });
        return cfTimeout;
    }

    /**
     * Send a one way message, without adding a completable future.
     *
     * @param ctx     The context to send the message under.
     * @param message The message to send.
     */
    public void sendMessage(ChannelHandlerContext ctx, CorfuMsg message) {
        ChannelHandlerContext outContext = context;
        if (ctx == null) {
            if (context == null) {
                // if the router's context is not set, return a failure
                log.warn("Attempting to send on a channel that is not ready.");
                return;
            }
            outContext = context;
        }
        // Get the next request ID.
        final long thisRequest = requestID.getAndIncrement();
        // Set the base fields for this message.
        message.setClientID(parameters.getClientId());
        message.setRequestID(thisRequest);
        message.setEpoch(epoch);
        // Write this message out on the channel.
        outContext.writeAndFlush(message, outContext.voidPromise());
        log.trace("Sent one-way message: {}", message);
    }


    /**
     * Send a netty message through this router, setting the fields in the outgoing message.
     *
     * @param ctx    Channel handler context to use.
     * @param inMsg  Incoming message to respond to.
     * @param outMsg Outgoing message.
     */
    public void sendResponseToServer(ChannelHandlerContext ctx, CorfuMsg inMsg, CorfuMsg outMsg) {
        outMsg.copyBaseFields(inMsg);
        outMsg.setEpoch(epoch);
        ctx.writeAndFlush(outMsg, ctx.voidPromise());
        log.trace("Sent response: {}", outMsg);
    }

    /**
     * Complete a given outstanding request with a completion value.
     *
     * @param requestId  The request to complete.
     * @param completion The value to complete the request with
     * @param <T>        The type of the completion.
     */
    public <T> void completeRequest(long requestId, T completion) {
        CompletableFuture<T> cf;
        if ((cf = (CompletableFuture<T>) outstandingRequests.get(requestId)) != null) {
            cf.complete(completion);
            outstandingRequests.remove(requestId);
        } else {
            log.warn("Attempted to complete request {}, but request not outstanding!", requestId);
        }
    }

    /**
     * Exceptionally complete a request with a given cause.
     *
     * @param requestID The request to complete.
     * @param cause     The cause to give for the exceptional completion.
     */
    public void completeExceptionally(long requestID, Throwable cause) {
        CompletableFuture cf;
        if ((cf = outstandingRequests.get(requestID)) != null) {
            cf.completeExceptionally(cause);
            outstandingRequests.remove(requestID);
        } else {
            log.warn("Attempted to exceptionally complete request {}, but request not outstanding!",
                requestID);
        }
    }

    /**
     * Validate the epoch of a CorfuMsg, and send a WRONG_EPOCH response if
     * the server is in the wrong epoch. Ignored if the message type is reset (which
     * is valid in any epoch).
     *
     * @param msg The incoming message to validate.
     * @param ctx The context of the channel handler.
     * @return True, if the epoch is correct, but false otherwise.
     */
    private boolean validateEpoch(CorfuMsg msg, ChannelHandlerContext ctx) {
        // Check if the message is intended for us. If not, drop the message.
        if (!msg.getClientID().equals(parameters.getClientId())) {
            log.warn("Incoming message intended for client {}, our id is {}, dropping!",
                msg.getClientID(), parameters.getClientId());
            return false;
        }
        // Check if the message is in the right epoch.
        if (!msg.getMsgType().ignoreEpoch && msg.getEpoch() != epoch) {
            log.trace("Incoming message with wrong epoch, got {}, expected {}, message was: {}",
                msg.getEpoch(), epoch, msg);
            /* If this message was pending a completion, complete it with an error. */
            completeExceptionally(msg.getRequestID(), new WrongEpochException(msg.getEpoch()));
            return false;
        }
        return true;
    }

    @Override
    protected void channelRead0(ChannelHandlerContext ctx, CorfuMsg m) throws Exception {
        try {
            // We get the handler for this message from the map
            IClient handler = handlerMap.get(m.getMsgType());
            if (handler == null) {
                // The message was unregistered, we are dropping it.
                log.warn("Received unregistered message {}, dropping", m);
            } else {
                if (validateEpoch(m, ctx)) {
                    // Route the message to the handler.
                    if (log.isTraceEnabled()) {
                        log.trace("Message routed to {}: {}",
                            handler.getClass().getSimpleName(), m);
                    }
                    handler.handleMessage(m, ctx);
                }
            }
        } catch (Exception e) {
            log.error("Exception during read!", e);
        }
    }

    @Override
    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) {
        log.error("Exception during channel handling.", cause);
        ctx.close();
    }

    @Override
    public void channelRegistered(ChannelHandlerContext ctx) throws Exception {
        context = ctx;
        log.debug("Registered new channel {}", ctx);
    }

    @Override
    public void channelUnregistered(ChannelHandlerContext ctx) throws Exception {
        super.channelUnregistered(ctx);
        context = null;
        log.debug("Unregistered channel {}", ctx);
    }

    @Override
    public void userEventTriggered(ChannelHandlerContext ctx, Object evt) throws Exception {
        if (evt.equals(ClientHandshakeEvent.CONNECTED)) {
            // Handshake successful. Complete the connection future to allow
            // clients to proceed.
            connectionFuture.complete(null);
        } else if (evt.equals(ClientHandshakeEvent.FAILED) && connectionFuture.isDone()) {
            // Handshake failed. If the current completion future is complete,
            // create a new one to unset it, causing future requests
            // to wait.
            connectionFuture = new CompletableFuture<>();
        }
    }

    // region Deprecated Methods
    // The methods below are deprecated and may be removed in a future release.
    /**
     * Creates a new NettyClientRouter connected to the specified endpoint.
     *
     * @param endpoint Endpoint to connectAsync to.
     * @deprecated Use {@link this#NettyClientRouter(NodeLocator, CorfuRuntimeParameters)}
     */
    @Deprecated
    public NettyClientRouter(String endpoint) {
        this(endpoint.split(":")[0], Integer.parseInt(endpoint.split(":")[1]));
    }

    /**
     * Creates a new NettyClientRouter connected to the specified host and port.
     *
     * @param host Host to connectAsync to.
     * @param port Port to connectAsync to.
     * @deprecated Use {@link this#NettyClientRouter(NodeLocator, CorfuRuntimeParameters)}
     */
    @Deprecated
    public NettyClientRouter(String host, Integer port) {
        this(NodeLocator.builder().host(host).port(port).build(),
            CorfuRuntimeParameters.builder().build());
    }

    /**
     * Creates a new NettyClientRouter connected to the specified host and port.
     *
     * @param host Host to connectAsync to.
     * @param port Port to connectAsync to.
     * @deprecated Use {@link this#NettyClientRouter(NodeLocator, CorfuRuntimeParameters)}
     */
    @Deprecated
    public NettyClientRouter(String host, Integer port, Boolean tls,
        String keyStore, String ksPasswordFile, String trustStore,
        String tsPasswordFile, Boolean saslPlainText, String usernameFile,
        String passwordFile) {
        this(NodeLocator.builder().host(host).port(port).build(),
            CorfuRuntimeParameters.builder()
                .tlsEnabled(tls)
                .keyStore(keyStore)
                .ksPasswordFile(ksPasswordFile)
                .trustStore(trustStore)
                .tsPasswordFile(tsPasswordFile)
                .saslPlainTextEnabled(saslPlainText)
                .usernameFile(usernameFile)
                .passwordFile(passwordFile)
                .build());
    }

    public NettyClientRouter(@Nonnull NodeLocator node,
        @Nonnull CorfuRuntimeParameters parameters) {
        this(node, parameters.getSocketType()
            .getGenerator().generate(Runtime.getRuntime().availableProcessors() * 2,
                new ThreadFactoryBuilder()
                    .setDaemon(true)
                    .setNameFormat(parameters.getNettyEventLoopThreadFormat())
                    .build()), parameters);
        shutdownEventLoop = true;
    }

    @Deprecated
    @Override
    public Integer getPort() {
        return node.getPort();
    }

    @Deprecated
    public String getHost() {
        return node.getHost();
    }
    // endregion
}
